third_party/WebKit/LayoutTests/fast/css/getComputedStyle/getComputedStyle-right-auto.html

<!DOCTYPE html>
<style>
  #container {
    width: 500px;
  }
  #container > div {
    background-color: red;
    position: relative;
    left: 5px;
    right: auto;
    top: 5px;
    bottom: 5px;
  }
</style>
<script src="../../../resources/js-test.js"></script>
<p>
  getComputedStyle should return minus the value of left as right when right is auto.
</p>
<div id="container">
  <div id=ltr>
    If direction of containing block is ltr, bottom and right should be negative
  </div>
  bottom and right should be negative
</div>
<div id="container" dir="rtl">
  <div id=rtl>
    If direction of containing block is rtl, bottom and right should be negative
  </div>
  bottom and right should be negative
</div>
<script>
  var ltrElement = document.getElementById("ltr");
  var rtlElement = document.getElementById("rtl");

  shouldBe("window.getComputedStyle(ltrElement).top", "'5px'");
  shouldBe("window.getComputedStyle(ltrElement).right", "'-5px'");
  shouldBe("window.getComputedStyle(ltrElement).bottom", "'-5px'");
  shouldBe("window.getComputedStyle(ltrElement).left", "'5px'");

  shouldBe("window.getComputedStyle(rtlElement).top", "'5px'");
  shouldBe("window.getComputedStyle(rtlElement).right", "'-5px'");
  shouldBe("window.getComputedStyle(rtlElement).bottom", "'-5px'");
  shouldBe("window.getComputedStyle(rtlElement).left", "'5px'");
</script>
